MAP-792 Add attribute and usage deletion tests to LocationResource
Integrate tests for the deletion of details for a location's attributes and usage to LocationResourceIntTest.kt. The updates add tests to verify whether the 'ROLE_MAINTAIN_LOCATIONS' role can successfully request to empty attributes and usage maps. The changes improve the test coverage for location modifications and ensure appropriate responses for specific role types.