MAP-801 Update UpsertLocationRequest fields and tests
The UpsertLocationRequest fields have been updated to make 'lastModifiedDate' optional and 'lastUpdatedBy' mandatory. Now, when 'lastModifiedDate' is not provided, the current time is used by default. Also, changes were made in the tests to reflect these updates and to handle the scenario of syncing a new location and updating it.