MAP-801 Update UpsertLocationRequest fields and tests (#34)
* MAP-801 Refactor attributes representation in Location objects

This commit includes changes to refactor how attributes are represented in 'Location' objects across various files. The attributes are now represented as a 'Set' of 'ResidentialAttributeValue' objects rather than a 'Map'. This simplifies the structure of the 'Location' entities.

* MAP-801 Update location entity creation and tests

The creation of ResidentialLocation and NonResidentialLocation entities has been updated to include attribute and usage information respectively. Correspondingly, the SyncAndMigrateResource integration tests have been updated to include attribute and usage information in the input test data.

* MAP-801 Update UpsertLocationRequest fields and tests

The UpsertLocationRequest fields have been updated to make 'lastModifiedDate' optional and 'lastUpdatedBy' mandatory. Now, when 'lastModifiedDate' is not provided, the current time is used by default. Also, changes were made in the tests to reflect these updates and to handle the scenario of syncing a new location and updating it.