MAP-614 Add location history tracking for changes and address exceptions
This commit introduces the ability to track changes to various location attributes so that the change history is preserved. Several new exceptions have been added in the case of certain conditions - for instance, already deactivated locations or existing ones. Furthermore, spelling corrections and attribute renaming have also been performed for better clarity and understanding of the code.