MAP-614 Add history flag to Location API methods (#41)
* Add toString, hashCode, equals methods and update logs

This commit implements toString, hashCode and equals methods for multiple classes. The log messages in LocationService class have been modified for improved traceability. Additional functionality has been included to track the history of changes made to Location attributes.

* MAP-614 Add history flag to Location API methods

This update includes an optional "includeHistory" boolean flag in the `getLocationById` and `getLocationByKey` methods in the Location service. This flag is defaulted to false and if set to true, will include historical changes of each location in the response. Relevant test cases are updated to accommodate this new feature.