MAP-896 Ignore duplicate history of change
Update Location and LocationHistory Model

The Location and LocationHistory models have been updated to include sorting features. In Location model, the field 'history' was changed from MutableList to SortedSet. Likewise, the Location_history model has been extended to implement Comparable interface. This will ensure that the history of locations are sorted naturally but also that duplicates will be ignored