MAP-896 Add deactivation details and update location indexes
Added deactivation details such as deactivatedDate, deactivationReason, and proposedReactivationDate to SyncAndMigrateResourceIntTest and UpsertLocationRequest files. Also created new indexes and dropped the old index in the V1_13__location_indexes.sql file.

In addition to updating the deactivated attribute from active to false for a location, deactivation details are now added to ensure better tracking and management of deactivated locations. Modifying the database indexes is intended to improve query performance related to location and location history.