MAP-896 Add deactivation details and update location indexes (#67)
* MAP-896 Add deactivation details and update location indexes

Added deactivation details such as deactivatedDate, deactivationReason, and proposedReactivationDate to SyncAndMigrateResourceIntTest and UpsertLocationRequest files. Also created new indexes and dropped the old index in the V1_13__location_indexes.sql file.

In addition to updating the deactivated attribute from active to false for a location, deactivation details are now added to ensure better tracking and management of deactivated locations. Modifying the database indexes is intended to improve query performance related to location and location history.

* Refactor deactivation process for prison locations

The commit implements improvements to the location deactivation process, specifically, a check is added to ensure that the deactivated date is provided when a deactivation request is made. On top of that, the code now leverages the `isDeactivated` function to verify if a location is already deactivated, streamlining the usage across different classes.