MAP-906 Improve Location capacity and attribute computation
Updated the way the system calculates capacity and attributes in ResidentialLocation to include the current cell, even if it is inactive. Renamed the method from isPermanentlyClosed() to isPermanentlyInactive() to better reflect its updated functionality. A new method isCurrentCellOrNotPermanentlyInactive() was introduced to simplify the computation. These changes were also reflected in the corresponding tests and Location DTO.