Refactor code and enhance logs for prisoner location checks (#92)
Removed the certification exception for checking cell capacity, as there is no longer a need to have this unnecessary check.
A refactor of the code was done to sort locations before checking for prisoners. Added logging for most functions that check for prisoners in a location. Some clean-ups on the test files were also made to improve readability.