MAP-872 Validate capacity and provide endpoints to change (#99)
* MAP-872 Validate capacity and provide endpoints to change

* MAP-872 Validate capacity and provide endpoints to change

* MAP-872 Add capacity and certification sync in update location request

The code now handles synchronization of capacity and certification in the cell update location request. This includes recording history of changes in capacity and certification details. Tests have been updated to account for these changes.

* MAP-872 Add capacity and certification sync in update location request

The code now handles synchronization of capacity and certification in the cell update location request. This includes recording history of changes in capacity and certification details. Tests have been updated to account for these changes.