MAP-1142 Add functionality to update deactivated location details
This update introduces functionality to update the details of a deactivated prison location. It allows changes to the deactivation reason, proposed reactivation date, and the planet FM reference. Also, this commit removed the `whenDeactivated` field from the Location table, and related entities and DTOs. Updates were made to the `V1_900__location_seed.sql` and the `LocationService.kt` files to incorporate these changes.