MAP-1101 Update accommodation type in specialist cells
This commit introduces an SQL migration script that updates the accommodation type field of specialist cells in the location table based on the residential housing type field. A case statement is used to determine the new accommodation type based on existing data. This change is applied only to active cells where the current accommodation type is 'OTHER_NON_RESIDENTIAL' and the residential housing type is not null.