MAP-978 Update handling of signed operation capacity
This commit includes the addition of a new function to publish changes related to signed operation capacity. It also refactors the handling of the saveSignedOperationalCapacity operation to have a separate code execution for new and existing records. Furthermore, the seed data for signed operation capacity in the database has been updated for testing purposes.