MAP-1173 Change deactivatedDate type to LocalDateTime
The type of "deactivatedDate" was changed from LocalDate to LocalDateTime in the Location class and related files. The test cases were updated to match this type change. Migration script was added to make the necessary change in the database. This change will provide more detailed deactivation timestamps including the time of the day.