MAP-1199 Update Gradle version and refactor CloudPlatformTask (#133)
This commit updates the Gradle version from 8.5 to 8.8 in gradle-wrapper.properties. It also simplifies the logic in CloudPlatformTask to retrieve a namespace prefix and environment. Additionally, the script URL referenced in 'gradlew' has been updated, the JVM target version is adjusted in the buildSrc 'build.gradle.kts' file and '.kotlin/' added to .gitignore to ignore module files generated in Kotlin.