MAP-1241 Refactor localName assignment in Location class (#138)
* MAP-1202 Update wheelchair accessible cells in DB

The commit updates specialist_cell records in the database where the specialist_cell_type is 'WHEELCHAIR_ACCESSIBLE', setting them to 'ACCESSIBLE_CELL'. It also removes any lingering records with the 'WHEELCHAIR_ACCESSIBLE' type to ensure data consistency.

* MAP-1241 Refactor localName assignment in Location class

An extracted method getDerivedLocalName has been introduced in the Location class to manage the assignment of localName. This refactoring avoids code duplication and enhances code readability. In the process, the method getActiveResidentialLocationsBelowThisLevel was also added.