MAP-1030 Update location groups using properties
This commit focuses on updating the filtering and retrieval of location groups using the properties defined within the relevant test. A new function `getCellLocationsForGroup` has been added to enable retrieval of cell locations for a specific group. Additional unit testing has been added to verify the functionality. Proper error handling has been introduced for cases when a group does not exist for a given prison ID.