MAP-1397 Add Legacy version of cell attributes for backward compatibility on CSC  (#177)
* MAP-1397 Add Legacy version of cell attributes for backward compatibility on CSC (for now!)

Extended the LocationOccupancyResource to include legacy attributes. The modifications also involve the addition of test cases to validate the new behaviour. Concurrently, the LocationService was updated to handle the creation of these new attributes.

* MAP-1397 Doc update

Extended the LocationOccupancyResource to include legacy attributes. The modifications also involve the addition of test cases to validate the new behaviour. Concurrently, the LocationService was updated to handle the creation of these new attributes.