MAP-1404 Add null checks before setting cell capacity and certification (#183)
The code was updated to verify if the cell capacity and certification are not null before setting them. This helps to avoid null pointer exceptions. If the capacity or certification is null, a new one is created with the provided parameters. Additionally, the `setCapacity` and `setCertification` methods have been added to the Certification and Capacity classes to encapsulate the setting logic.