MAP-1450 Disable Jira and Slack notifications for train deployment (#204)
* Update Prisoner Search API port from 8080 to 8090

Changed the `API_BASE_URL_PRISONER_SEARCH` and the corresponding `containerPort` for the Prisoner Search API service to port 8090. This ensures consistency and avoids port conflicts within the `values-train.yaml` configuration.

* Update docker images to latest in values-train.yaml

Changed the wiremock and localstack container images to always use the latest version. This ensures we are using the most recent updates and features in these containers.

* MAP-1450 Disable Jira and Slack notifications for train deployment

Jira updates and Slack notifications have been turned off for the "train" environment in the CircleCI configuration. This change simplifies the deployment process and avoids unnecessary notifications.