MAP-1561 Add endpoint to update location's local name (#230)
* MAP-1561 Add endpoint to update location's local name

Implemented a new PUT endpoint to change the local name of a location. Updated necessary service methods, DTOs, and test cases to support this.

* Remove unused import in LocationResidentialResourceTest

Eliminated the import of `Location` from `LocationResidentialResourceTest.kt` as it was not being used in the file. This helps in keeping the codebase clean and improves code readability.